Refactor XPath predicate evaluation into dedicated translation unit
Moved predicate comparison logic from xpath_evaluator.cpp to a new
xpath_evaluator_predicates.cpp file to improve code organisation and
modularity.

Changes:
- Created xpath_evaluator_predicates.cpp with 5 core predicate functions:
  * node_set_string_value() - Node-set to string conversion
  * node_set_number_value() - Node-set to number conversion
  * promote_value_comparison_operand() - Value normalisation
  * compare_xpath_values() - Equality comparison (=, !=, eq, ne)
  * compare_xpath_relational() - Relational comparison (<, >, <=, >=, lt, gt, le, ge)

- Updated xpath_evaluator_detail.h to declare predicate functions for
  shared access between translation units

- Removed 209 lines of predicate code from xpath_evaluator.cpp

- Updated CMakeLists.txt to compile new predicates file

- Fixed forward declarations in xpath_value.h to match schema_types.h
  (added [[nodiscard]] and noexcept specifiers)

All predicate functions consume shared numeric utilities (numeric_equal,
numeric_compare) from xpath_evaluator_detail.h for consistent behaviour.

Test Results: 100% pass rate (14/14 XPath test suites)
