+++ src/xml/xpath/xpath_evaluator_detail.h
@@ -3,6 +3,8 @@
 #include "xpath_value.h"
 
 #include <memory>
+#include <optional>
+#include <string>
 
 enum class RelationalOperator {
    LESS,
@@ -11,8 +13,18 @@ enum class RelationalOperator {
    GREATER_OR_EQUAL
 };
 
+// Schema and type system helpers
 std::shared_ptr<xml::schema::SchemaTypeDescriptor> schema_descriptor_for_value(const XPathValue &Value);
 bool should_compare_as_boolean(const XPathValue &Left, const XPathValue &Right);
 bool should_compare_as_numeric(const XPathValue &Left, const XPathValue &Right);
+
+// Numeric comparison utilities
 bool numeric_equal(double Left, double Right);
 bool numeric_compare(double Left, double Right, RelationalOperator Operation);
+
+// Predicate value extraction and comparison (implemented in xpath_evaluator_predicates.cpp)
+std::string node_set_string_value(const XPathValue &Value, size_t Index);
+double node_set_number_value(const XPathValue &Value, size_t Index);
+std::optional<XPathValue> promote_value_comparison_operand(const XPathValue &Value);
+bool compare_xpath_values(const XPathValue &LeftValue, const XPathValue &RightValue);
+bool compare_xpath_relational(const XPathValue &LeftValue, const XPathValue &RightValue, RelationalOperator Operation);
